feat: Separated install process + autoloader. First approach
