+++ lib/EasyREST.php
@@ -9,6 +9,9 @@
  * @copyright Doofinder
  * @license   GPLv3
  */
+
+namespace PrestaShop\Module\Doofinder\Lib;
+
 if (!defined('_PS_VERSION_')) {
     exit;
 }
