+++ lib/doofinder_layer_api.php
@@ -12,7 +12,7 @@
  * @copyright Doofinder
  * @license   GPLv3
  */
-require_once _PS_MODULE_DIR_ . 'doofinder/lib/EasyREST.php';
+use PrestaShop\Module\Doofinder\Lib\EasyREST;
 
 if (!defined('_PS_VERSION_')) {
     exit;
